rename go module paths
