+++ services/el_clients/geth/specular/cmd/geth/config.go
@@ -42,11 +42,11 @@ import (
 	"github.com/ethereum/go-ethereum/node"
 	"github.com/ethereum/go-ethereum/params"
 	"github.com/naoina/toml"
-	specularUtils "github.com/specularl2/specular/clients/geth/specular/cmd/utils"
-	entry "github.com/specularl2/specular/clients/geth/specular/entry/geth"
-	"github.com/specularl2/specular/clients/geth/specular/internal/ethapi"
-	"github.com/specularl2/specular/clients/geth/specular/internal/flags"
-	"github.com/specularl2/specular/clients/geth/specular/rollup"
+	specularUtils "github.com/specularl2/specular/services/el_clients/geth/specular/cmd/utils"
+	entry "github.com/specularl2/specular/services/el_clients/geth/specular/entry/geth"
+	"github.com/specularl2/specular/services/el_clients/geth/specular/internal/ethapi"
+	"github.com/specularl2/specular/services/el_clients/geth/specular/internal/flags"
+	"github.com/specularl2/specular/services/el_clients/geth/specular/rollup"
 )
 
 var (
