+++ services/el_clients/geth/specular/cmd/geth/main.go
@@ -40,10 +40,10 @@ import (
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/ethereum/go-ethereum/metrics"
 	"github.com/ethereum/go-ethereum/node"
-	specularUtils "github.com/specularl2/specular/clients/geth/specular/cmd/utils"
-	"github.com/specularl2/specular/clients/geth/specular/internal/debug"
-	"github.com/specularl2/specular/clients/geth/specular/internal/ethapi"
-	"github.com/specularl2/specular/clients/geth/specular/internal/flags"
+	specularUtils "github.com/specularl2/specular/services/el_clients/geth/specular/cmd/utils"
+	"github.com/specularl2/specular/services/el_clients/geth/specular/internal/debug"
+	"github.com/specularl2/specular/services/el_clients/geth/specular/internal/ethapi"
+	"github.com/specularl2/specular/services/el_clients/geth/specular/internal/flags"
 
 	// Force-load the tracer engines to trigger registration
 	_ "github.com/ethereum/go-ethereum/eth/tracers/js"
