+++ services/el_clients/geth/specular/entry/geth/fee_hook.go
@@ -9,7 +9,7 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/core/vm"
-	"github.com/specularl2/specular/clients/geth/specular/rollup/utils/log"
+	"github.com/specularl2/specular/services/el_clients/geth/specular/rollup/utils/log"
 )
 
 // TODO: maybe move these constants to the rollup config or a separate file
