+++ services/el_clients/geth/specular/rollup/rpc/eth/txmgr/tx_mgr.go
@@ -14,7 +14,7 @@ import (
 	"github.com/ethereum/go-ethereum/core"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/log"
-	"github.com/specularl2/specular/clients/geth/specular/rollup/utils/fmt"
+	"github.com/specularl2/specular/services/el_clients/geth/specular/rollup/utils/fmt"
 )
 
 // Geth defaults the priceBump to 10
