+++ services/el_clients/geth/specular/rollup/services/sequencer/batcher.go
@@ -13,7 +13,7 @@ import (
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/ethereum/go-ethereum/params"
 	"github.com/ethereum/go-ethereum/trie"
-	"github.com/specularl2/specular/clients/geth/specular/rollup/services"
+	"github.com/specularl2/specular/services/el_clients/geth/specular/rollup/services"
 )
 
 // Batcher assumes exclusive control of underlying blockchain, i.e.
